Problem: CI servers don't support multiple processes at the same time
Solution: Create the cucumber_feature_runner which wraps the logic used
in the cucumber_runner so it can be used programmatically.

With the cucumber_feature_runner a cucumber tests runner can be
constructed the runs the feature runner and the steps runner in one
processes which will then work in CI servers.